Week_11/java-at-a-conference/src/main/java/Person.java
public class Person {
  private String firstName;
  private String lastName;
  private String email;

  public Person(String firstName, String lastName, String email) {
    this.firstName = firstName;
    this.lastName = lastName;
    this.email = email;
  }

  public String getFirstName() {
    return this.firstName;
  }

  public String getLastName() {
    return this.lastName;
  }

  public String getEmail() {
    return this.email;
  }

  public String getName() { return this.firstName + " " + this.lastName; }
}
